refactor: unify sync/async get_redis helpers and update API exports
