+++ packages/kiarina-lib-redis/src/kiarina/lib/redis/asyncio.py
@@ -3,10 +3,16 @@ from importlib import import_module
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from ._async.registry import get_redis
-    from .settings import settings_manager
+    from ._async.helpers.get_redis import get_redis
+    from ._settings import RedisSettings, settings_manager
 
-__all__ = ["get_redis", "settings_manager"]
+__all__ = [
+    # ._async.helpers
+    "get_redis",
+    # ._settings
+    "RedisSettings",
+    "settings_manager",
+]
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
 
@@ -16,8 +22,11 @@ def __getattr__(name: str) -> object:
         raise AttributeError(f"module '{__name__}' has no attribute '{name}'")
 
     module_map = {
-        "get_redis": "._async.registry",
-        "settings_manager": ".settings",
+        # ._async.helpers
+        "get_redis": "._async.helpers.get_redis",
+        # ._settings
+        "RedisSettings": "._settings",
+        "settings_manager": "._settings",
     }
 
     parent = __name__.rsplit(".", 1)[0]
